Небольшие архитектурные изменения
