+++ src/main/java/ru/reboot/organizer/routing/handlers/button/menu/link/LinkPlatformButtonHandler.java
@@ -1,4 +1,4 @@
-package ru.reboot.organizer.routing.handlers.button;
+package ru.reboot.organizer.routing.handlers.button.menu.link;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -10,22 +10,26 @@ import ru.reboot.organizer.utils.dev.UserScreens;
 
 @Component
 @RequiredArgsConstructor
-public class EnterCodeButtonHandler implements ButtonHandler {
+public class LinkPlatformButtonHandler implements ButtonHandler {
     private final SessionManager sessionManager;
 
     @Override
     public ButtonType getHandledButton() {
-        return ButtonType.ENTER_LINK_CODE;
+        return ButtonType.LINK_PLATFORM;
     }
 
     @Override
     public UnifiedResponse handleButton(Long userId) {
-        sessionManager.setUserScreen(userId, UserScreens.PLATFORM_LINK_INPUT);
+        sessionManager.setUserScreen(userId, UserScreens.LINK_PLATFORM_MENU);
 
         return UnifiedResponse.builder()
-                .text("Введите 6-значный код привязки:")
+                .text("Подключение платформы")
                 .row()
-                .button("Отмена", ButtonType.MAIN_MENU)
+                .button("Сгенерировать код", ButtonType.GENERATE_LINK_CODE)
+                .row()
+                .button("Ввести код", ButtonType.ENTER_LINK_CODE)
+                .row()
+                .button("В главное меню", ButtonType.MAIN_MENU)
                 .build();
     }
 }
